--- GraphDB/GraphDB.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33110.190
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GraphDB", "GraphDB\GraphDB.vcxproj", "{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Debug|x64.ActiveCfg = Debug|x64
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Debug|x64.Build.0 = Debug|x64
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Debug|x86.ActiveCfg = Debug|Win32
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Debug|x86.Build.0 = Debug|Win32
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Release|x64.ActiveCfg = Release|x64
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Release|x64.Build.0 = Release|x64
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Release|x86.ActiveCfg = Release|Win32
-		{FC20B545-ABFA-4BEA-AE3F-4CAB874FE812}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9B5B0D89-D9AD-4ECD-9627-F3C3A9B22E20}
-	EndGlobalSection
-EndGlobal
