--- GraphDB/GraphDB/GraphDB.vcxproj.filters
@@ -1,14 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <ClCompile Include="DataCluster.cpp" />
-    <ClCompile Include="FileReader.cpp" />
-    <ClCompile Include="GraphDB.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="DataCluster.h" />
-    <ClInclude Include="FileReader.h" />
-    <ClInclude Include="Graph.h" />
-    <ClInclude Include="Node.h" />
-  </ItemGroup>
-</Project>
