--- GraphDB/GraphDB/Tree.cpp
@@ -1,5 +1,0 @@
-#include "Tree.h"
-int main(){
-    int n=5;
-    cout << n;
-}
