--- ProjectTesting/ProjectTesting.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33122.133
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ProjectTesting", "ProjectTesting\ProjectTesting.vcxproj", "{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Debug|x64.ActiveCfg = Debug|x64
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Debug|x64.Build.0 = Debug|x64
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Debug|x86.ActiveCfg = Debug|Win32
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Debug|x86.Build.0 = Debug|Win32
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Release|x64.ActiveCfg = Release|x64
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Release|x64.Build.0 = Release|x64
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Release|x86.ActiveCfg = Release|Win32
-		{5B0601D7-5BB8-4FAC-B70F-C0427FB8907B}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B139885E-322C-4339-A407-E245ACC8222F}
-	EndGlobalSection
-EndGlobal
