--- ProjectTesting/ProjectTesting/pch.h
@@ -1,7 +1,0 @@
-//
-// pch.h
-//
-
-#pragma once
-
-#include "gtest/gtest.h"
