+++ TestingData/TestData.csv
@@ -1,0 +1,6 @@
+﻿Name,school,age,district,Faculty
+oshan,BC,23,Gampaha,ENG
+nehra,SV,22,Mathara,ENG
+ashen,DV,23,Gampaha,ENG
+navindu,EX,22,Colombo,ENG
+oshan,RC,21,Kagalla,ENG
