+++ Tree.cpp
@@ -1,0 +1,5 @@
+#include "Tree.h"
+int main(){
+    int n=5;
+    cout << n;
+}
